+++ timeago/example/main.dart
@@ -13,11 +13,16 @@ main() async {
   timeago.setLocaleMessages('dv_short', timeago.DvShortMessages());
   timeago.setLocaleMessages('fr', timeago.FrMessages());
   timeago.setLocaleMessages('fr_short', timeago.FrShortMessages());
+  timeago.setLocaleMessages('gr', timeago.GrMessages());
+  timeago.setLocaleMessages('gr_short', timeago.GrShortMessages());
   timeago.setLocaleMessages('ca', timeago.CaMessages());
   timeago.setLocaleMessages('ca_short', timeago.CaShortMessages());
+  timeago.setLocaleMessages('cs', timeago.CsMessages());
+  timeago.setLocaleMessages('cs_short', timeago.CsShortMessages());
   timeago.setLocaleMessages('ja', timeago.JaMessages());
   timeago.setLocaleMessages('km', timeago.KmMessages());
   timeago.setLocaleMessages('km_short', timeago.KmShortMessages());
+  timeago.setLocaleMessages('ko', timeago.KoMessages());
   timeago.setLocaleMessages('id', timeago.IdMessages());
   timeago.setLocaleMessages('pt_BR', timeago.PtBrMessages());
   timeago.setLocaleMessages('pt_BR_short', timeago.PtBrShortMessages());
@@ -44,6 +49,7 @@ main() async {
   timeago.setLocaleMessages('ko', timeago.KoMessages());
   timeago.setLocaleMessages('vi', timeago.ViMessages());
   timeago.setLocaleMessages('vi_short', timeago.ViShortMessages());
+  timeago.setLocaleMessages('tr', timeago.TrMessages());
   timeago.setLocaleMessages('ta', timeago.TaMessages());
   timeago.setLocaleMessages('ro', timeago.RoMessages());
   timeago.setLocaleMessages('ro_short', timeago.RoShortMessages());