+++ timeago/lib/src/messages/da_messages.dart
@@ -1,5 +1,6 @@
 import 'package:timeago/src/messages/lookupmessages.dart';
 
+/// Danish Messages
 class DaMessages implements LookupMessages {
   @override
   String prefixAgo() => '';
@@ -35,6 +36,7 @@ class DaMessages implements LookupMessages {
   String wordSeparator() => ' ';
 }
 
+/// Danish short Messages
 class DaShortMessages implements LookupMessages {
   @override
   String prefixAgo() => '';