+++ timeago/lib/src/timeago.dart
@@ -2,6 +2,8 @@ import 'package:timeago/src/messages/en_messages.dart';
 import 'package:timeago/src/messages/es_messages.dart';
 import 'package:timeago/src/messages/lookupmessages.dart';
 
+String _default = 'en';
+
 Map<String, LookupMessages> _lookupMessagesMap = {
   'en': EnMessages(),
   'en_short': EnShortMessages(),
@@ -9,12 +11,26 @@ Map<String, LookupMessages> _lookupMessagesMap = {
   'es_short': EsShortMessages(),
 };
 
+/// Sets the default [locale]. By default it is `en`.
+///
+/// Example
+/// ```
+/// setLocaleMessages('fr', FrMessages());
+/// setDefaultLocale('fr');
+/// ```
+void setDefaultLocale(String locale) {
+  assert(locale != null, '[locale] must not be null');
+  assert(_lookupMessagesMap.containsKey(locale),
+      '[locale] must be a registered locale');
+  _default = locale;
+}
+
 /// Sets a [locale] with the provided [lookupMessages] to be available when
 /// using the [format] function.
 ///
 /// Example:
 /// ```dart
-/// setLocaleMessages('fr', FrMessages())
+/// setLocaleMessages('fr', FrMessages());
 /// ```
 ///
 /// If you want to define locale message implement [LookupMessages] interface
@@ -36,7 +52,7 @@ void setLocaleMessages(String locale, LookupMessages lookupMessages) {
 ///   5 minutes from now in 'en' locale will display as "5 minutes from now"
 String format(DateTime date,
     {String locale, DateTime clock, bool allowFromNow}) {
-  final _locale = locale ?? 'en';
+  final _locale = locale ?? _default;
   final _allowFromNow = allowFromNow ?? false;
   final messages = _lookupMessagesMap[_locale] ?? EnMessages();
   final _clock = clock ?? DateTime.now();