+++ timeago_flutter/CHANGELOG.md
@@ -1,3 +1,13 @@
-## [0.0.1] - TODO: Add release date.
+## [0.3.0] - Improvements to flutter widget to support auto refresh @kranfix.
 
-* TODO: Describe initial release.
+- Creating a TimerRefreshWidget as a base for TimerRefresh
+- Making packge compatible with timeago all 2.x.x
+
+## [0.2.0] - Improvements to flutter widget to support auto refresh @kranfix.
+
+- Refactoring for using less code for timeago_flutter
+- New widget TimerRefresh
+
+## [0.1.0] - Initial release.
+
+- Timeago widget